lstm-spam-api/requirements.txt
numpy==1.19.5
pandas==1.3.3
scikit-learn==0.24.2
tensorflow==2.6.0
fastapi==0.110.1
cassandra-driver==3.29.1